Move class definitions in `elf_utils.h` to a compilation unit
Prepare the file to add new functions that can be imported from
different compilation units.
